fix(design-tokens): move constants to TokenTable level
